requirements.txt
git+git://github.com/Rapptz/discord.py@rewrite#egg=discord-py
dropbox
attrdict
pyyaml